requirements.txt
pyspark==3.3.0
mysql-connector-python==8.0.21
bson==0.5.10